Created simple grid map
Created a function that draws a grid map for the floor and fills the grids for the walls for a nice and understandable map to program translation. Also made the function that draws a simple dot at the coordinates of the player on the map. Next step is casting our first singular ray from the player position in the direction of the player with working wall colision.