Made the first ray cast
Made the first simple ray cast, just a simple line drawing that is checking after every pixel if the new coordinates are in a wall cell, it will be upgraded into the fully functional DDA algorith soon. Also made the posibility to rotate and move the player on the map with the ray adjusting at the same time. I need to look into a way of pressing multiple keys at one time for like example holding W and A at the same time to move diagonally or rotating and moving in the same time. ATTENTION: the code doesnt compile at the moment, for correct compilation delete the "t_player" struct from the cub3d_h and put back the variables into the "t_game" struct.