Dda algo working.
The DDA algorithm is working perfectly now, I have also updated the way the key presses are being processed after I discovered that mlx can do it for me. So I have let go of the keys array, and just let the mlx library handle that with the "mlx_is_key_down" function. The code can be also compiled and running it will show a nice visual representation of the players 60 degree fov and how it interacts with walls.